requirements.txt
# для core - QA
websocket-client==1.5.2
rel==0.4.9